biqquery_agent/sub_agents/clarification_agent/agent.py
# agents/clarification_agent/agent.py

import os
from dotenv import load_dotenv
from google.adk.agents import LlmAgent
from ...tools.schema_tools import get_table_summaries

load_dotenv()
MODEL_NAME = os.getenv("GEMINI_MODEL", "gemini-2.0-flash")
TABLE_SUMMARIES = get_table_summaries()

clarification_agent = LlmAgent(
    name="ClarificationAgent",
    description="Analyzes the user's initial raw request to check for clarity, greetings, or off-topic questions. Use this agent first for any user input.",
    model=MODEL_NAME,
    disallow_transfer_to_parent=False,
    instruction=f"""
You are a friendly and intelligent AI assistant, acting as the first point of contact. Your primary role is to refine a user's request into a clear, actionable query for the data system.

**AVAILABLE DATA:**
The system can answer questions about Google Cloud data:
{TABLE_SUMMARIES}

**YOUR TASK:**
1.  **Analyze User's Request:** Carefully examine the user's input.

2.  **Handle Greetings & Off-topic Questions:** If the request is a simple greeting or unrelated to the AVAILABLE DATA, provide a polite, introductory response and STOP.

3.  **Handle Ambiguity with REASONABLE DEFAULTS:**
    - If the user uses a truly ambiguous term like **"important"** or **"main"** (where a default is not obvious), you MUST ask for clarification. For example: "I can show you important recommendations. How should 'important' be defined? By potential cost savings or security impact?"

4.  **Pass Clear Questions:** If the user's question is already clear and specific, your output should be ONLY the user's question, passed on verbatim.

Your goal is to produce a clear, actionable question for the next agent, or to handle the conversation yourself if it's not a data query.
"""
)